requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.0.0
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.3
black==23.9.1
blinker==1.8.2
blis==0.7.11
boto3==1.34.74
botocore==1.34.144
bs4==0.0.2
catalogue==2.0.10
certifi==2024.2.2
cffi==1.15.1
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
cryptography==41.0.3
cymem==2.0.8
dataclasses-json==0.6.7
distlib==0.3.8
distro==1.9.0
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
et-xmlfile==1.1.0
exceptiongroup==1.2.2
faiss-cpu==1.8.0
fastapi==0.110.0
fastapi-versioning==0.10.0
filelock==3.15.4
flake8==6.1.0
Flask==3.0.2
Flask-Cors==4.0.0
flatbuffers==24.3.25
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
humanfriendly==10.0
identify==2.6.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.7
langchain-community==0.2.7
langchain-core==0.2.18
langchain-openai==0.1.16
langchain-text-splitters==0.2.2
langcodes==3.4.0
langsmith==0.1.85
language_data==1.2.0
marisa-trie==1.2.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mccabe==0.7.0
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.3
nltk==3.8.1
nodeenv==1.9.1
numpy==1.26.4
onnxruntime==1.18.1
openai==1.35.13
opencv-python==4.10.0.84
openpyxl==3.1.2
orjson==3.10.6
packaging==23.2
passlib==1.7.4
pathspec==0.11.2
pillow==10.4.0
platformdirs==3.10.0
pluggy==1.3.0
pre-commit==3.7.0
preshed==3.0.9
protobuf==5.27.2
py-mon==2.0.5
pyasn1==0.5.0
pyclipper==1.3.0.post5
pycodestyle==2.11.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pyflakes==3.1.0
PyJWT==2.8.0
pymongo==4.6.3
pypdf==4.3.0
pytest==7.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
PyYAML==6.0.1
rapidocr-onnxruntime==1.3.15
regex==2024.5.15
requests==2.31.0
rsa==4.9
s3transfer==0.10.2
shapely==2.0.5
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.31
srsly==2.4.8
starlette==0.36.3
sympy==1.13.0
tenacity==8.3.0
thinc==8.2.5
tiktoken==0.7.0
tomli==2.0.1
tqdm==4.66.4
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.11.0rc1
urllib3==2.2.2
uvicorn==0.30.1
virtualenv==20.26.3
wasabi==1.1.3
watchdog==4.0.1
weasel==0.3.4
Werkzeug==3.0.3
xlrd==2.0.1
yarl==1.9.4
